move tests to test_nn
